use centralized redis instances (aws elasticache) and don't run updaters on every worker node
