+++ ci/ansible/roles/deploy/defaults/main/executables.yml
@@ -4,7 +4,6 @@ scripts_path: "/usr/local/bin"
 
 dbs_update: "almalinux_dbs_update"
 mirrors_list_update: "almalinux_mirrors_list_update"
-mirrors_list_update_pid: "{{ source_path }}/{{ mirrors_list_update }}.pid"
 mirrors_service: "almalinux_mirrors_service"
 utilities_service: "almalinux_mirrors_utilities"
 ...
