--- ci/ansible/roles/deploy/defaults/main/redis.yml
@@ -1,3 +1,0 @@
----
-redis_conf: "/etc/redis/redis.conf"
-...
