--- ci/ansible/roles/deploy/tasks/redis_configuration.yml
@@ -1,45 +1,0 @@
----
-- name: Disable persistent storage for Redis
-  lineinfile:
-    path: "{{ redis_conf }}"
-    regexp: '^appendonly yes$'
-    line: 'appendonly no'
-    state: present
-    insertbefore: BOF
-
-- name: Disable RDB snapshots for Redis
-  lineinfile:
-    path: "{{ redis_conf }}"
-    regexp: '^save .*'
-    state: absent
-    insertbefore: BOF
-
-- name: Disable RDB snapshots for Redis
-  lineinfile:
-    path: "{{ redis_conf }}"
-    line: 'save ""'
-    state: present
-    insertbefore: BOF
-
-- name: Enable redis socket
-  lineinfile:
-    path: "{{ redis_conf }}"
-    regexp: '^unixsocket'
-    line: unixsocket {{ redis_socket_path }}
-    state: present
-    insertbefore: BOF
-
-- name: Set redis socket perms
-  lineinfile:
-    path: "{{ redis_conf }}"
-    regex: '^unixsocketperm'
-    line: unixsocketperm 777
-    state: present
-    insertbefore: BOF
-
-- name: Start Redis
-  systemd:
-    enabled: yes
-    masked: no
-    state: restarted
-    name: redis
