+++ ci/ansible/roles/deploy/templates/almalinux_mirrors_list_update.service.j2
@@ -2,10 +2,10 @@
 Description={{ description }} Service
 Wants=network-online.target
 After=network.target network-online.target
-After=almalinux_mirrors_utilities.service
 
 [Service]
 User={{ service_user }}
 Group={{ service_user }}
 Type=oneshot
 ExecStart={{ scripts_path }}/{{ mirrors_list_update }}.sh
+TimeoutStartSec=10min
