--- ci/ansible/roles/deploy/templates/almalinux_mirrors_utilities.service.j2
@@ -1,28 +1,0 @@
-[Unit]
-Description=AlmaLinux mirrors service (Gunicorn backend)
-After=syslog.target
-After=almalinux_mirrors_service.service
-
-
-[Service]
-User={{ service_user }}
-Group={{ service_user }}
-Type=simple
-Restart=on-failure
-RestartSec=5
-TimeoutStopSec=10
-WorkingDirectory={{ source_path }}/src/backend
-ExecStartPre={{ scripts_path }}/{{ utilities_service }}.sh
-ExecStart={{ source_path }}/bin/gunicorn -c {{ source_path }}/src/backend/service.py
-
-LimitAS=infinity
-LimitRSS=infinity
-LimitCORE=infinity
-LimitNOFILE=65536
-LimitNPROC=infinity
-LimitSIGPENDING=infinity
-LimitSIGPENDING=infinity
-LimitMSGQUEUE=infinity
-
-[Install]
-WantedBy=multi-user.target
