--- ci/ansible/roles/deploy/templates/almalinux_mirrors_utilities.sh.j2
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-if [[ -f "{{ mirrors_list_update_pid }}" ]]; then
-    rm -f "{{ mirrors_list_update_pid }}"
-fi
-exit 0
