+++ ci/ansible/roles/deploy/templates/gunicorn.py.j2
@@ -18,8 +18,7 @@ raw_env = [
     'DEPLOY_ENVIRONMENT={{ config_root }}/deploy_environment.var',
     'TEST_IP_ADDRESS={{ test_ip_address }}',
     'AUTH_KEY={{ auth_key }}',
-    'REDIS_SOCKET=/var/run/redis/redis.sock',
-    'MIRRORS_UPDATE_PID={{ mirrors_list_update_pid }}',
+    'REDIS_URI=redis://mirrorlistservice-ecredis.tpowt8.ng.0001.use1.cache.amazonaws.com:6379',
 ]
 accesslog = '-'
 access_log_format = '"%({X-Forwarded-For}i)s" %(l)s %(u)s %(t)s "%(r)s" %(s)s %(b)s "%(f)s" "%(a)s"'
