--- ci/ansible/roles/deploy/templates/service.py.j2
@@ -1,23 +1,0 @@
-wsgi_app = 'utilities:app'
-bind = '[::1]:{{ utilities_port }}'
-workers = 4
-user = '{{ service_user }}'
-group = '{{ service_user }}'
-raw_env = [
-    'GEOIP_PATH={{ config_root }}/mirrors/geoip_db.mmdb',
-    'ASN_PATH={{ config_root }}/mirrors/asn_db.mmdb',
-    'SQLITE_PATH={{ config_root }}/data/mirrors.db',
-    'SENTRY_DSN={{ sentry_dsn }}',
-    'PYTHONPATH={{ source_path }}/src/backend/',
-    'CONFIG_ROOT={{ config_root }}',
-    'SOURCE_PATH={{ source_path }}',
-    'DEPLOY_ENVIRONMENT={{ config_root }}/deploy_environment.var',
-    'TEST_IP_ADDRESS={{ test_ip_address }}',
-    'AUTH_KEY={{ auth_key }}',
-    'REDIS_SOCKET=/var/run/redis/redis.sock',
-    'MIRRORS_UPDATE_PID={{ mirrors_list_update_pid }}',
-]
-accesslog = '-'
-access_log_format = '"%({X-Forwarded-For}i)s" %(l)s %(u)s %(t)s "%(r)s" %(s)s %(b)s "%(f)s" "%(a)s"'
-errorlog = '-'
-timeout = 0
