+++ src/backend/api/handlers.py
@@ -232,187 +232,6 @@ def _get_nearest_mirrors(
     return suitable_mirrors
 
 
-async def update_mirrors_handler() -> str:
-
-    main_config = get_config(
-        logger=logger,
-        path_to_config=SERVICE_CONFIG_PATH,
-        path_to_json_schema=SERVICE_CONFIG_JSON_SCHEMA_DIR_PATH,
-    )
-    mirrors_dir = os.path.join(
-        os.getenv('CONFIG_ROOT'),
-        'mirrors/updates',
-        main_config.mirrors_dir,
-    )
-    all_mirrors = get_mirrors_info(
-        mirrors_dir=mirrors_dir,
-        logger=logger,
-        main_config=main_config,
-        path_to_json_schema=MIRROR_CONFIG_JSON_SCHEMA_DIR_PATH,
-    )
-
-    pid_file_path = Path(os.getenv('MIRRORS_UPDATE_PID'))
-    if pid_file_path.exists():
-        return 'Update is already running'
-    time1 = time.time()
-    message = 'Update of the mirrors list is finished at "%s"'
-    try:
-        logger.info('Update of the mirrors list is started')
-        pid_file_path.write_text(str(os.getpid()))
-        step = 100
-        mirror_check_sem = asyncio.Semaphore(step)
-        mirrors_len = len(all_mirrors)
-        async with mirror_check_sem, MirrorProcessor(
-            logger=logger,
-        ) as mirror_processor:  # type: MirrorProcessor
-            mirror_iso_uris = mirror_processor.get_mirror_iso_uris(
-                versions=set(main_config.versions) -
-                set(main_config.duplicated_versions),
-                arches=main_config.arches
-            )
-            subnets = await get_aws_subnets(
-                http_session=mirror_processor.client,
-            )
-            subnets.update(
-                await get_azure_subnets(
-                    http_session=mirror_processor.client,
-                ),
-            )
-            for i in range(0, mirrors_len, step):
-                next_slice = min(i + step, mirrors_len)
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_subnets_for_cloud_mirror(
-                            subnets=subnets,
-                            mirror_info=mirror_info,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.cloud_type
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_ip_for_mirror(
-                            mirror_info=mirror_info,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_iso_url(
-                            mirror_info=mirror_info,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_status_of_mirror(
-                            main_config=main_config,
-                            mirror_info=mirror_info,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.ip not in ('Unknown', None)
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_ipv6_support_of_mirror(
-                            mirror_info=mirror_info,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.ip not in ('Unknown', None)
-                    and not mirror_info.private
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_geo_and_location_data_from_db(
-                            mirror_info=mirror_info)
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.status in ('ok', 'expired')
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_mirror_have_full_iso_set(
-                            mirror_info=mirror_info,
-                            mirror_iso_uris=mirror_iso_uris,
-                        )
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.status in ('ok', 'expired')
-                    and mirror_info.ip not in ('Unknown', None)
-                    and not mirror_info.private
-                    and mirror_info.cloud_type in ('', None)
-                ))
-                await asyncio.gather(*(
-                    asyncio.ensure_future(
-                        mirror_processor.set_location_data_from_online_service(
-                            mirror_info=mirror_info)
-                    ) for mirror_info in all_mirrors[i:next_slice]
-                    if mirror_info.status in ('ok', 'expired')
-                ))
-        with session_scope() as db_session:
-            db_session.query(Mirror).delete()
-            db_session.query(Url).delete()
-            db_session.query(Subnet).delete()
-            for mirror_info in all_mirrors:
-                urls_to_create = [
-                    Url(
-                        url=url,
-                        type=url_type,
-                    ) for url_type, url in mirror_info.urls.items()
-                ]
-                db_session.add_all(urls_to_create)
-                mirror_to_create = Mirror(
-                    name=mirror_info.name,
-                    continent=mirror_info.geolocation.continent,
-                    country=mirror_info.geolocation.country,
-                    state=mirror_info.geolocation.state_province,
-                    city=mirror_info.geolocation.city,
-                    ip=mirror_info.ip,
-                    ipv6=mirror_info.ipv6,
-                    latitude=mirror_info.location.latitude,
-                    longitude=mirror_info.location.longitude,
-                    status=mirror_info.status,
-                    update_frequency=dateparser.parse(
-                        mirror_info.update_frequency
-                    ),
-                    sponsor_name=mirror_info.sponsor_name,
-                    mirror_url=mirror_info.mirror_url,
-                    iso_url=mirror_info.iso_url,
-                    sponsor_url=mirror_info.sponsor_url,
-                    email=mirror_info.email,
-                    cloud_type=mirror_info.cloud_type,
-                    cloud_region=mirror_info.cloud_region,
-                    urls=urls_to_create,
-                    private=mirror_info.private,
-                    monopoly=mirror_info.monopoly,
-                    asn=','.join(mirror_info.asn),
-                    has_full_iso_set=mirror_info.has_full_iso_set,
-                )
-                if mirror_info.subnets:
-                    subnets_to_create = [
-                        Subnet(
-                            subnet=subnet,
-                        ) for subnet in mirror_info.subnets
-                    ]
-                    db_session.add_all(subnets_to_create)
-                    mirror_to_create.subnets = subnets_to_create
-                db_session.add(mirror_to_create)
-        # update all mirrors lists in the Redis cache
-        for args in itertools.product(
-                (True, False),
-                repeat=len(signature(get_all_mirrors_db).parameters),
-        ):
-            cache_key = _generate_redis_key_for_the_mirrors_list(*args)
-            cache.delete(cache_key)
-            get_all_mirrors_db(*args)
-    finally:
-        logger.info(
-            'Update of the mirrors list is finished at "%s"',
-            time.time() - time1,
-        )
-        if pid_file_path.exists():
-            os.remove(pid_file_path)
-    return message % (time.time() - time1)
-
-
 def get_all_mirrors(
         get_working_mirrors: bool = False,
         get_expired_mirrors: bool = False,
